requirements.txt
anyio==3.6.1
boto3~=1.24.42
botocore~=1.27.42
click
fastapi==0.78.0
h11==0.13.0
idna==3.3
jmespath
pydantic==1.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
s3transfer
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.2
requests
pytest==6.2.2
gunicorn==20.1.0
virtualenv==20.4.3
supervisor==4.2.2
toml==0.10.2
visitor==0.1.3
Werkzeug==1.0.1
pillow~=9.2.0
face_recognition
opencv-python~=4.6.0.66
numpy~=1.23.1
tensorflow~=2.9.1
keras~=2.9.0